+++ groupvote/src/main/java/com/supergroup/groupvote/GroupVoteRepository.java
@@ -2,6 +2,6 @@ package com.supergroup.groupvote;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface GroupVoteRepository extends CrudRepository<GroupVote, Long> {
+public interface GroupVoteRepository extends CrudRepository<GroupVote, Integer> {
 
 }
